Refactor based on utils discussion
