+++ server/middleware.go
@@ -1,42 +1,140 @@
 package server
 
 import (
-	"github.com/tozny/e3db-clients-go/authClient"
-	"github.com/tozny/utils-go"
+	"bytes"
+	"context"
+	"encoding/json"
+	"io/ioutil"
 	"log"
 	"net/http"
 	"strings"
+
+	"github.com/tozny/e3db-clients-go/authClient"
+)
+
+const (
+	// ToznyClientIDHeader is the headers key containing a client ID
+	ToznyClientIDHeader = "X-TOZNY-CLIENT-ID"
+	// ToznyOpenAuthenticationTokenHeader is the header key contaning a Tot
+	ToznyOpenAuthenticationTokenHeader = "X-TOZNY-TOT"
+	// HealthCheckPathSuffix is a centrally defined health check path.
+	HealthCheckPathSuffix = "/healthcheck"
+	// ServiceCheckPathSuffix is a centrally defined service check path.
+	ServiceCheckPathSuffix = "/servicecheck"
+)
+
+var (
+	// DefaultCORSHeaders is a full set of CORS headers for use in the CORS middleware
+	DefaultCORSHeaders = []http.Header{
+		// https://developer.mozilla.org/en-US/docs/Web/HTTP/CORS#The_HTTP_response_headers
+		map[string][]string{
+			"Access-Control-Allow-Origin":      []string{"*"},
+			"Access-Control-Allow-Methods":     []string{"*, GET, POST, DELETE, PUT, OPTIONS, HEAD"}, // Because to Firefox * does not mean all.
+			"Access-Control-Allow-Headers":     []string{"Authorization, Content-Type, *"},           // Because to Firefox * does not mean all.
+			"Access-Control-Allow-Credentials": []string{"true"},
+			"Access-Control-Max-Age":           []string{"86400"},
+		},
+	}
 )
 
 // Middleware is a function that decorates an http.Handler
+//
+// The decorator function can determine whether to pass the request on to the
+// next handler in the chain by calling the ServeHTTP method on the handler. If
+// the middleware should pass additional information along with the request,
+// context is available on the request object. Add a value to the context.
 type Middleware func(handler http.Handler) http.Handler
 
-// ApplyMiddleware creates a final http Handler with a slice of middleware functions applied
-func ApplyMiddleware(handlers []Middleware, final http.Handler) http.Handler {
+// ApplyMiddleware returns the final handler wrapped in all passed middleware handlers.
+func ApplyMiddleware(final http.Handler, handlers ...Middleware) http.Handler {
 	for _, handler := range handlers {
 		final = handler(final)
 	}
 	return final
 }
 
-// JSONLoggingMiddleware converts the utils JsonLogger into Middleware
+// JSONLoggingMiddleware wraps an HTTP handler and logs
+// the request and de-serialized JSON body.
 func JSONLoggingMiddleware(logger *log.Logger) Middleware {
 	return func(h http.Handler) http.Handler {
-		return utils.JsonLoggingHandler(h, logger)
+		return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+			var requestBody interface{}
+			json.NewDecoder(r.Body).Decode(&requestBody)
+			logger.Print(map[string]interface{}{
+				"request_method":    r.Method,
+				"request_uri":       r.RequestURI,
+				"requester_address": r.RemoteAddr,
+				"requester_host":    r.Host,
+				"request_body":      requestBody,
+			})
+			// Repopulate body with the data read
+			jsonBytes := new(bytes.Buffer)
+			json.NewEncoder(jsonBytes).Encode(requestBody)
+			r.Body = ioutil.NopCloser(jsonBytes)
+			h.ServeHTTP(w, r)
+		})
 	}
 }
 
-// CORSMiddleware converts the utils CORSHandler into Middleware
+// CORSMiddleware provides http middleware for allowing cross origin requests by
+// decorating the request with the provided CORS headers and returning default 200 OK for any options requests
 func CORSMiddleware(corsHeaders []http.Header) Middleware {
 	return func(h http.Handler) http.Handler {
-		return utils.CORSHandler(h, corsHeaders)
+		return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+			for _, corsHeader := range corsHeaders {
+				for key, values := range corsHeader {
+					for _, value := range values {
+						w.Header().Set(key, value)
+					}
+				}
+			}
+			switch r.Method {
+			case http.MethodOptions:
+				HandleOptionsRequest(w)
+				return
+			}
+			h.ServeHTTP(w, r)
+		})
 	}
 }
 
-// AuthMiddleware converts utils E3dbAuthHandler into Middleware
+// AuthMiddleware provides http middleware for enforcing requests as coming from e3db
+// authenticated entities (either external or internal clients) for any request with a path
+// not ending in `HealthCheckPathSuffix` or `ServiceCheckPathSuffix`
 func AuthMiddleware(auth authClient.E3dbAuthClient, privateService bool, logger *log.Logger) Middleware {
 	return func(h http.Handler) http.Handler {
-		return utils.E3dbAuthHandler(h, auth, privateService, logger)
+		return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+			// Check to see if this request is a health or service check requests
+			requestPath := r.URL.Path
+			isMonitoringRequest := strings.HasSuffix(requestPath, HealthCheckPathSuffix) || strings.HasSuffix(requestPath, ServiceCheckPathSuffix)
+			if isMonitoringRequest {
+				// NoOp authentication, continue processing request
+				h.ServeHTTP(w, r)
+				return
+			}
+			token, err := ExtractBearerToken(r)
+			if err != nil {
+				logger.Printf("E3dbAuthHandler: error extracting bearer token %s\n", err)
+				HandleError(w, http.StatusUnauthorized, err)
+				return
+			}
+			ctx := context.Background()
+			validateParams := authClient.ValidateTokenRequest{
+				Token:    token,
+				Internal: privateService,
+			}
+			validateTokenResponse, err := auth.ValidateToken(ctx, validateParams)
+			if err != nil || !validateTokenResponse.Valid {
+				logger.Printf("E3dbAuthHandler: error validating token %s %+v\n", err, validateTokenResponse)
+				HandleError(w, http.StatusUnauthorized, ErrorInvalidAuthToken)
+				return
+			}
+			// Add the clients id and token to the request headers
+			r.Header.Set(ToznyClientIDHeader, validateTokenResponse.ClientId)
+			r.Header.Set(ToznyOpenAuthenticationTokenHeader, token)
+			// Authenticated, continue processing request
+			h.ServeHTTP(w, r)
+		})
 	}
 }
